requirements.txt
hyperliquid-python-sdk==0.18.0
eth-account==0.10.0
python-dotenv==1.0.0
